Add inputattach support for the EETI eGalaxTouch serial touchscreen driver v2.
Add inputattach support for the EETI eGalaxTouch serial touchscreen driver.

Also add support for the Hampshire touchscreen driver.

v2:

Made the eGalaxTouch code silent.

Made the Hampshire support unconditional. Reasoning:
The value of SERIO_HAMPSHIRE is smaller than SERIO_PS2MULT,
so it was added earlier to the kernel. inputattach has
unconditional support for the latter, so it should cause
no problem for the former, either.

Signed-off-by: Böszörményi Zoltán <zboszor@pr.hu>
Signed-off-by: Stephen Kitt <steve@sk2.org>